refactor(frontend): split god files exceeding 500 lines
